okieflat-site/requirements.txt
Flask<1
PyYAML>3